Задача 4. С перехвата ИИ хода человека. Другой вариант подсчета победы
